+++ Elevator/Movimientos.java
@@ -2,8 +2,8 @@ package Elevator;
 
 public interface Movimientos {
     // public void activar();
-    public void mover(int duracionMovimiento);
+    public void mover(double duracionMovimiento);
 
-    public void parada(int tiempoEspera);
+    public void parada(double tiempoEspera);
 
 }
